+++ examples/security-center/main.tf
@@ -1,8 +1,8 @@
 terraform {
-  required_version = "~>0.12.15"
+  required_version = "> 0.12.26"
 
   required_providers {
-    azurerm = "~> 1.36"
+    azurerm = ">= 2.31"
   }
 }