+++ modules/security-center/outputs.tf
@@ -1,6 +1,6 @@
-output "security_center_pricing_id" {
-  description = "The Security Center subscription pricing ID."
-  value       = azurerm_security_center_subscription_pricing.pricing.id
+output "security_center_pricing_ids" {
+  description = "The Security Center subscription pricing IDs list."
+  value       = [for p in azurerm_security_center_subscription_pricing.pricing : p.id]
 }
 
 output "security_center_contact_id" {