+++ modules/security-center/variables.tf
@@ -7,6 +7,12 @@ variable "security_center_pricing_tier" {
   default     = "Free"
 }
 
+variable "security_center_pricing_resource_types" {
+  description = "List of resource type to apply the Azure Security Center pricing tier. Possible values are AppServices, ContainerRegistry, KeyVaults, KubernetesService, SqlServers, SqlServerVirtualMachines, StorageAccounts, and VirtualMachines. Source: https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/security_center_subscription_pricing#resource_type"
+  type        = list(string)
+  default     = ["AppServices", "ContainerRegistry", "KeyVaults", "KubernetesService", "SqlServers", "SqlServerVirtualMachines", "StorageAccounts", "VirtualMachines"]
+}
+
 variable "security_center_contact_email" {
   description = "The email of the Security Center Contact. https://www.terraform.io/docs/providers/azurerm/r/security_center_contact.html#email"
   type        = string