+++ outputs.tf
@@ -1,9 +1,9 @@
 ###############################
 # Security Center outputs
 ###############################
-output "security_center_pricing_id" {
-  description = "The Security Center subscription pricing ID."
-  value       = module.security-center.security_center_pricing_id
+output "security_center_pricing_ids" {
+  description = "The Security Center subscription pricing IDs list."
+  value       = module.security-center.security_center_pricing_ids
 }
 
 output "security_center_contact_id" {